Separate keyboard code into multiple files
We want to support multiple keyboard backends (Wayland and Xcb), but
in order to do this we first need to separate the code to make it
more manageable.

Issue: #440
