+++ src/greeter/CMakeLists.txt
@@ -8,12 +8,14 @@ set(GREETER_SOURCES
     ${CMAKE_SOURCE_DIR}/src/common/SocketWriter.cpp
     GreeterApp.cpp
     GreeterProxy.cpp
+    KeyboardLayout.cpp
     KeyboardModel.cpp
     ScreenModel.cpp
     SessionModel.cpp
     ThemeConfig.cpp
     ThemeMetadata.cpp
     UserModel.cpp
+    XcbKeyboardBackend.cpp
 )
 
 add_executable(sddm-greeter ${GREETER_SOURCES})
